backend/requirements/base.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
django-unfold>=0.40,<1.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
Pillow>=10.4,<11.0
redis>=5.0,<6.0
drf-spectacular>=0.29,<1.0
